Reduce view render logging we consider excessive, in production
As I try to monitor the app in heroku, I am finding current logs pretty hard to deal with due to excessive verbosity. A more drastic change to logging could be taken ala #656, but without wanting to take time to figure that out and configure it, for now we can just turn off view/render logging (which is also part of what lograge or semantic_logs would do), which will get us a lot of the way there to less awful logs, or at least we can see how far it gets us.