Update dependencies that can be updated
Simply ran `bundle update` to update all dependencies that can be update within dependency restrictions in our Gemfile and dependency tree.

Good to do this periodically. Note that we are intentionally (in our Gemfile) locking Blacklight to an older release for now, becuase of our experience with frequent backwards incompatibilities in new releases. At some point we can spend time on vetting new BL version and possibly updating to it, but we're not choosing that in this PR.

Note also that browse-everything is unfortunately preventing updates of some dependencies, in confusing ways/reasons, which is also a separate issue. Here we just bundle update to update what we an.