also update browse-everything and dependencies
The latest release of browse-everything finally allows us to upgrade to it, and still use Rails 6.0 and sprockets 4. This also finally allows us to upgrade to thor 1.x, faraday 1.0.0, and google_drive 3.x

Note we are still stuck on faraday 1.0.0 unable to upgrade to 1.0.1 becuase of too restrictive dropbox_api dependency.
https://github.com/samvera/browse-everything/issues/361