heroku REDIS_URL can be rediss too, set ssl_params for that one too just in case
On eddie's app it's giving a rediss: url in REDIS_URL, and no REDIS_TLS_URL, as opposed to mine where it has a rediss: URL in REDIS_TLS_URL and redis: in REDIS_URL. Maybe because eddie is on heroku-20 stacck already? Unclear, heroku is being a bit sloppy. But assume that REDIS_URL can have a rediss too, we need to be sure to set ssl_params for that one too.