GH Actions CI, try apt update before apt-get
Builds were failing trying to install libpulse0, one of the dependencies of the things we tried to install with apt-get. Don't know why it used to work and stopped.

But googling the error got us to this:
https://github.community/t/libpulse0-fails-to-fetch-with-apt-repo-broken/18214/3

Which suggested you use `apt update` first. So we do that, except the `apt-get update` variant, cause we're using `apt-get`. Phew.