git pre-commit hook included in repo to try and catch AWS keys before commit
To use, you need to set a git config value in your local copy, to tell git to use hooks that came with the repo. Or you could copy it or whatever.

Anyway, this won't be foolproof, but sometimes helps.

Kind of wrote my own mashing together a few others, in comments.

Ref #433