alt tag on some rights statements improves accessibility somewhat
I am seeing why the rightsstatement usage notes tell you not to do what we did, makes it hard to make sure it's accessible. We hadn't found this guidance when we implemented it like this, and it DOES look a lot better... for now we're sticking with it. https://rightsstatements.org/en/documentation/usage_guidelines.html