ignore anchors with = in them for jumping to tab
Becuase we're going to use anchors with = in them for other purposes, such as timecodes.

Yes, this means there's no way to jump to a tag and a timecode. but it seems nice to leave the tab anchors as plain old HTML style fragment identifiers. But in the future we could expand to allow like #fragment=tabAnchor&t=timecode