make attr_json a direct dependency
It is a dependency of kithe; but since it's not a direct dependency, it doesn't get automatically loaded on boot. So this can cause weird auto-loading issues if our app tries to use it before it's loaded code from kithe that causes attr_json to get required. Since we use it ourselves directly, we should just express it as a dependency.

I may also respond to this in kithe by making kithe require attr_json on boot.

This should fix the weird auto-loading problem from #1014 among others.