add tom-select.js filtered multi-select for large-element association select inputs
tom-select JS and CSS added via npm/webpacker, in the "admin" pack that's added only to admin layout.

For the moment, we are NOT doing any AJAX loading here (although tom-select supports it) -- we are still loading all select options in delivered HTML, just using tom-select for a much nicer UI for it.

Used for:

* Work: select collection
* Work/Oral History: select linked interviewer profile
* Work/Oral History: select linked interviewee biography