preserve tab selection in url as eg #tab=ohTranscript
We add tab selection to URL #anchor on selection, and on page load switch to that tab. This lets tab selection be reflected in the URL, so when you bookmark you get current tab. Oral History audio navbar is the use case, although it's active on all bootstrap tabs.

We were already doing this prior to this PR, but putting the ID of the selected tab in teh anchor as the entire anchor, eg "#ohTranscript"

we now use a key/value format instead, like "#tab=ohTranscript".  This gives us room to store more than one thing in anchor, in particular upcoming planned feature for linking specifically to a timecode, which might look like #tab=ohTranscript&t=1212" where 1212 is a number of seconds.