Redirect some existing legacy OH microsite search result links
Only bothering with the main searches, not fielded searches. Will catch things like the search form on main sciencehistory.org website, if it's left as is. Final decision hasn't been made wrt what to do with OH search on main site; but it's so cheap/easy to do this redirect, let's do it anyway, just for completeness in redirects. Either it'll be used or it won't, doesn't hurt anything by being there.

Maybe i just think setting up redirects in the routing file is fun.

Ref #1123