Don't try to restart/reload solr on deploy
Our existing restart/reload code wouldn't work with searchstax -- and we're trying to switch to searchstax in production as part of emergency recovery measures.

We do have a rake task to send current solr config to searchstax solr. But we're not trying to wire it up to cap deploy in this PR. Instead, we just disable the function.

The rake task is `scihist:solr_cloud:sync_configset`.  Once we have production EC2 using searchstax solr, if you know you have a solr config change that needs to be sync'd, you should be able to ssh to an EC2, say, jobs server, and from there run `./bin/rake scihist:solr_cloud:sync_configset`.