capistrano deploy will sync solr configset to solrcloud
Before we had searchstax/solrcloud, we sync'd solr config on every deploy using a different mechanism. Now we do it by running the rake task we wrote to sync using solr cloud APIs, it will be run on a deployment server so it has appropriate environment (solr url, passwords, etc).