use custom logger formatter to avoid logging things heroku is adding in anyway
or things that are irrelevant/unintelligible on heroku. Help reduce our log data, for things like papertrail that have paid plans baesd on log data.

In fact, all we're adding with the formatter is log level (info, debug, etc) -- maybe we don't even need that, and could use no log formatter at all, but I like seeing it and being able to search on it, so we'll leave it for now.