Don't create a download_ style derivative, if it would be larger than original
These are meant to downscale to create smaller downloads; there is no reason to do it or offer it to the user if it would be a LARGER size than original. Which for some download sizes, we noticed happening in some cases. The DOWNLOAD_FULL derivative (jpg at *same size* as the original, but in jpg form) will still be created, regardless of size.

Our existing code for displaying download options already will smartly not display an option if the derivative isn't there. So by refraining from CREATING the derivative, it also won't show up in download options, bingo bango bongo.

So this one line is much of the implementation of #1246.

But to close that one completely, we also need to go through existing data and remove download derivatives that are larger than their originals.