rake scihist:checks_fixity:complete_overdue
Refactor/add method to FixityReport so we can get the actual list of assets, not just the count, that are displayed in dashboard as problematic. FixityReport is a little bit messy, all those existing methods probably should have had a name making it clear they were returning a count. Oh well.

Then add rake task that uses this to just... check them. Since planning on doing from a scheduled job, no progress bar. Just outputs to log if we found any stale ones and checked em... where we will see that it happened only if we go looking for it. Does it need any other kind of alerting? Not sure.