Clearer error message when WorkPDFCreator doesn't have any images
Avoid making a failed attempt to call pdfunite. Have a clearer error message in our logs for what actually the failure was.

Clean up the tempfile while we're at it. (Not a big deal, as ruby will clean it up eventually, but we can help the GC).

Ref: #1283

see also #1168 for one context in which this condition can occur; there could be others though at least hypothetically, like bots caching a URL for a work that used to be PDF-able but isn't anymore?