update overridden shrine method to proper method signature
These two are treated equivalently in ruby 2.7 and prior, but in ruby 3.0 with it's changes to separation between keyword arguments and hash data, they aren't the same at all. Shrine updated this signature at some point to be ruby 3 compatible, we need to update to match similarly prior to going to ruby 3 (discovered in running tests under ruby 3), might as well do it now.