Fix attempt to do Jquery-style 'on' delegation without Jquery
Previous attempt failed when the click was a *child* of the delegations selector. Turns out this is a bit trickier than totally naive. But not that much. Copied an implementation from StackOverflow. It's a bit hacky, but it works, to resolve the problem in #1017

Still don't totally understand how to keep my JS files and imports organized well, not sure this is quite right... but it works.