Try to robots.txt most catalog search results
Specifically those including any facet limits.

Trick is getting the syntax right to disallow what we want, and NOT accidentally disallowing collection landing pages, which are actually high-priority to be indexed! I think this will work.

My ideal first try solution would not be this, but instead `nofollow` on the facet limit links, but waiting for a Blacklight PR-merge-release for that.

Ref: #1306