Stop adding extra # in URL manipulation
Every time we tried to edit the URL on opening or closing viewer, we were adding an extra "#" in, because location.hash already includes a leading "#".

This resulted in URLs in the browser that ended up looking like eg https://digital.sciencehistory.org/works/dqao8vj###tab=transcription

Which actually broke the "auto-select tab on page load" functionality from working too!

This fixes that.