Set and handle Solr timeout
Set a maximum value we're willing to wait for Solr -- tricky to figure out right key for this, but as of present RSolr, it's that `read_timeout` value in blacklight.yml

But now what happens when it's raised? Blacklight will currently turn it into a Blacklight::Exceptions::InvalidRequest, the same as an http 4xx or 5xx from Solr. And with any of those, Blacklight will keep the exception from being caught by Honeybadger etc, but instead re-display search page with an (inaccurate) “Sorry, I don’t understand your search.” message.

By overriding handle_request_error in CatalogController, we turn off this behavior. Just allow the error to raise, resulting in our generic error page, logging, and honeybadger tracking. AND we want a timeout to be distinguishable, look at the Blacklight::Exceptions::InvalidRequest error and if it was originally a Faraday::TimeoutError, actually raise THAT.

For research leading to this patch: https://bibwild.wordpress.com/2021/10/11/setting-and-handling-solr-timeouts-in-blacklight/