error handling pages should handle ANY formats
With just a plain text response I guess. If error happens in some format we haven't explicitly enumerated, still do something other than an error. Without this, the error handling itself ends up raising a ActionController::UnknownFormat, which triggers Rails failsafe exception handling. And results in eg a 500 in what should be a 404.

This triggered #1362