Fix accidental duplicate id=description_tab attribute
This was supposed to be transcription_tab, but accidentally copied and pasted description_tab.

Accessibility tool caught it as: "Duplicate id attribute value "description-tab" found on the web page."

In addition to being just illegal HTML, it can be a problem for assistive technology associating the tab with it's content.

Ref WCAG work at #565