Hide rights statement icon from assistive technology consistently
The text next to it is sufficient, the icon can be hidden.

We were doing that somewhat inconsistently, we switch to doing it consistently, with blank `alt` attriute AND aria-hidden=true on all of them -- both large format and compact format.

From getting advice on code4lib slack, I _think_ this is the right thing to do in all of these circumstances. Doing my best!

Ref WCAG work at #565