Add assistive tech labels for navigational thumbnails in viewers
It's not entirely clear what to do with these, even after talking to a friendly person on code4lib slack who knows more than me.

This is my best attempt to hopefully make it better. Those thumbnails used for navigation should have assistive-tech-readable labels like "Image 1", "Image 2" -- becuase that's pretty much all we can say about them, and it's better than nothing at least.

We supply this as both an `alt` attribute and an `aria-label` attribute, becuase of concenrs that existing `role=button` might make assistive tech ignore the alt and want the aria-label. It may depend on the tech. The role=button may or may not actually be a good idea in the first place, but it's our best guess, since these images ARE serving as buttons!  At some point with this stuff I start getting into things where the right answer is not clear to me, depsite limited attempts to research. Doing my best.

Ref WCAG work #565