remove duplicate ids search-submit-header
One of the accessibility checkers caught it, in some cases we had more than one id=search-submit-header on page, which is illegal HTML.

It doesn't look like we're USING this id for anything, I don't think it needs to be there at all. Removing from all except main search form.

Ref WCAG work #565