On the back-end catch bad date range params that trigger Solr error
The front-end is already preventing it via #1422.

But someone can still craft a URL that would trigger a Solr error, and show up in our error logs, ala #1415. So we catch that on the back-end too. Becuase it shoudln't be possible for a human using our actual front-end to do this, we can provide the same sort of unfriendly plain-text http 400 error message we do for other similar param validation, it should not be possible for a human using our front-end to see it.