Point to existing footnote text in aria-describedby instead of repeating it
If we need the aria-describedby to point to footnote text -- we already have that text in the footnotes section, and don't need to repeat it again at the footnote reference. Among other things, this will help us later deal with the duplicate ID problems in #1411, by having fewer ID's involved.

I'm not sure if we actually need the aria-describedby or not, but we'll leave it there, figuring we had a good reason to put it in, so it's still there, pointing to in fact the same text, after this change.