remove aria-hidden
It's actually semi-illegal here, you shouldn't put aria-hidden on something that wraps an interactive element, unless the interactive element has tabindex=-1.

https://dequeuniversity.com/rules/axe/4.1/aria-hidden-focus

And in this case, we want the footnote link to be keyboard navigable, we don't actually want it to be tabindex=-1. I'm not sure what the aria-hidden was intended to accomplish, removing it.