remove anchor on cancel URL
The anchor was preventing the cancel url from working sometimes -- after a validation error, the cancel url would be the same as the URL you were already looking at in browser, and when it included a #fragment, the browser would not actually reload the page on clicking it, as intended. oops.

Also this was the wrong fragment anchor for our new tab system anyway. AND the tab it was trying to highlight is default tab for this page anyway, not really necessary.

Just remove it!