add current user to honeybadger context
Useful so we can check with the person doing the ingest when an error happens in ingest. Although actually won't work if the error happened in a bg job as it often does, I now realize. Becuase there's no logged in user there. Oh well, can't hurt might help in some situations.