Query searches id field too
By moving it out of ransack so we can just supply our own logic for handling the UUID. We cast it so it just turns into NULL if it's not a valid UUID format, and then OR it together with the other text fields -- the text fields use ILIKE case-insenstive like, but the UUID field just uses equality.

Ransack is convenient when you are on it's path, but when fighting it to do somethin weird... we end up using it less and less, resulting in a weird hybrid. Oh well!