Allow test Solr to be started externally and left running
Spec setup uses a simple net 'ping' sort of code to see if something is connectable on host/port solr is expected. If so, it assumes test solr is running.

If not, it still tries to start it automatically if needed, and then shuts it down after tests.

But since solr is slow to start up and shut down, you probably just want to leave it running.

Additionally,this means that profiling and timing of our specs won't include the slow solr startup in it's measurements, which is good.

Github CI has been changed to explicitly start up solr too.

We might eventually remove altogether the solr auto-startup code, it's a bit fragile.  But not here yet.

Ref #1482