Protect /facet action from bad params too
The /facet action is used for fetching the list of facets for the "more facets" paginating display. In main catalog controller, and in collection_show controllers that sub-class it.
We were protecting the /catalog action itself from various kind of bad parameters -- making sure they got caught and got a pretty error message, instead of being un-rescued 500s.

We need to do the same thing for the /facet action, which takes the same limit params to constrain the search.

Re: tests, we could and perhaps "should" be repeating ALL of our bad param tests for each context it can be in -- catalog controller vs collection_show; /catalog action vs /facet action. Four repeats of each test. We hypothetically could be doing that using rspec shared tests feature, although it makes it harder the way we are doing the tests, which are a bit disorganized.  We aren't doing those repeats of all tests. We just added one more test for the actual problem encountered in #1510. I did do it TDD though.

Ref #1510