even top and bottom whitespace in deai footer
It was annoying me that the top and bottom padding in the footer had ended up somehow uneven. Fixed, with maybe a hair more breathing room too.

Yeah, I think all of this use of overrides and especially \!important means our css is a bit out of control, but so it goes.