When trying to start up solr, if the collection name already exists just keep going
This came up because when we switched how we're doing CI Solr setup, we ended up having GH Actions cache a copy of solr that already had the collection. I think. Not totally sure why this wasn't always a problem.

Anyway, now if the collection already exists, instead of bombing when we try to create it, we just continue on happy.

Not totally sure if this will end up causing additional problems. This stuff starts getting messy, it's true.