No need for exemptions from webmock control for Solr
We use WebMock to "mock" HTTP requests in tests, but also to *prevent* the app from making external HTTP requests in tests -- we don't want tests depending on external remote HTTP sites, generally, and WebMock can hook into ruby to raise if something tries.

Our webmock disallow net connect command already allowed connections to localhost though -- this is standard setup to get Capybara tests to work, and just generally.

So we don't need to turn off the webmock disallow for our solr-using tests. Which is what the code had been doing previously, I guess I thought it was needed so the app could get to solr for solr-using tests. But it's not, our disallow already had an exemption for localhost, and we want to leave the rest of the disallow in place.

Our tests all still pass fine when we remove the special "allow net connect, then run the test, then disallow it again" around tests tagged :solr. So remove the unncesssary fiddling.