Need to pass args as real keywords in ruby 3
We were passing a hash to something that wanted keyword arguments. That doesn't work in ruby 3.0 anymore -- we are dynamically constructing the keyword args from a hash, so we need to use the `**` operator to convert them to actual keyword arguments.

Caused an error in production even though tests didn't catch, because tests didn't exersize this Env code that only is used in development and test environments, for differnet shrine setup. Oops, so it goes.

Ref #1539