Patch Shrine S3 for tagging_directive: "REPLACE" on copy
Necessary for our configuration for automated content-type tagging.

PR'd and accepted to shrine at https://github.com/shrinerb/shrine/pull/569/

But hasn't been released in shrine yet. So we need to patch in a fixed version of the #copy method (the smallest unit we can patch).

Wrapped in sane_patch to raise and remind us to remove when we get an upgraded shrine that has this fix in it.