change CombinedAudioDerivativeCreator to call ffmpeg with url args
No need to download them locally as tempfiles if we can access them remotely as URLs, say an s3 signed url. ffmpeg can use the URLs! Should be somewhat more efficient timewise. Maybe more RAM-efficient too? Not sure.

We also change the logic to use new duration metadata, instead of previous code that re-extracted duration.

Note all existing tests pass, this is just a refactor hooray. EXCEPT the number of significant digits on duration changes, now we're only to ms level of granularity, before it was nanoseconds. which is fine.