update caniuse-lite db
Noticed this output in the webpack JS build log:

```
Browserslist: caniuse-lite is outdated. Please run:
npx browserslist@latest --update-db

Why you should do it regularly:
https://github.com/browserslist/browserslist#browsers-data-updating
```

So I ran: `npx browserlist@latest --update-db`

It changed the yarnl.lock as so.

Why do we even have caniuse-lite? It's a dependency of autoprefixer, which is an indirect dependency of rails webpacker. (Which is kind of end-of-lifed, and at some point we should switch to a more maintained less hacky Rails JS pipeline. See #748