Generate longer expirationf for s3 urls used as ffmpeg input'
Possibly relevant to #1639. A default signed S3 URL is good for 15 minutes. We feed these URLs to ffmpeg to create a combined derivative. What if it takes longer than 15 minutes to create, and S3 starts refusing? You'd think we'd get an error message. But maybe not.

Just in case this is part of the problem, it doesn't hurt to make these S3 URLs we use as ffmpeg input last a lot longer. How about 48 hours.