Some refactoring of S3ConsoleURI, now with tests
We had been passing in http(s) URIs, and extracting bucket name from *.s3.amazonaws.com hostname. This will stop working if we switch to configuring a CDN like CloudFront in the shrine storage though -- in that the https URI generated won't have the bucketname in it.

We do not fully fix this problem here -- it's hard to fix in some of hte fixity/orphan checking paths, where the URL to an object of concern is all that's stored in the DB, and counted upon to do double-duty as encapsulating bucket name for Console URI generation AND a URL the file can actually be accessed at. However, we start setting the ground for solving it, and using it in a clearer way going forward.

* Separate the logic for getting bucket name from URI out from logic for generating console URI
* Add new initializer method for creating directly from a shrine UploadedFile object, in a way that doens't rely on guessing bucket from hostname. Or directly from bucket/path if you have it.
* Add some tests for the first time.

If we cannot get an S3 Console URI, we just return `nil` -- this is a bit weird and might make things harder to debug, but is there to fail gracefully for this "extra add-on" functionality -- especially becuase in a test and sometimes dev environmetn we don't always run with S3 storage, we want things to fail gracefully in those cases, not raise.

Going forward, it's always better to use an s3:// URI or a shrine UploadedFile object, not an https uri!