create HLS video routine respects our kithe backgrounding directives for derivatives
Yes, this code is getting very esoteric, sorry! but let's make the HLS creation we added on in a callback respect the same kithe promotion directives that apply to other derivatives. So if code directs kithe to disable derivative creation; or do it in the foreground instead of backgroudn -- this applies to HLS creation too.