Fix problems with HLS derivatives un-createable in test
If we try to run our HLS derivative creations job with a video original that is not in a Shrine S3 storage -- it raises. In test, our storages are not S3. (The raise here may or may not be a good idea, it does lead to some headaches and confusing situations. Perhaps we should configure our HLS creation to be by default not happen in test, and/or make it silently log failure instead of raise if it can't. But this PR doesn't do that).

Some past PR's resulted in a situation where this problem became visible, in a way that wasn't caught by PR builds -- becuase it resulted from interaction of several seperate PRs.  Test code may have been triggering HLS creation that failed before too... but the failures could be hidden/invisible, recent PR's to make things more consistent and predictable may have made them show up and fail tests.

So we use some kind of cryptic code to prevent HLS creation in those tests where it was failing. This *whole thing* has ended up rahter too byzantine/confusing for me, but for now it's what we've got.

This should get tests green again.