Add explicit `matrix` dependency for ruby 3.1
The prawn gem uses matrix; as of ruby 3.1 it needs to be declared explicitly. There isn't a prawn release that does that yet, although it's been fixed in prawn master. We can work around that by adding an explicit top-level dependency.

This didn't cause a problem in development (unless/until you tried to use a prawn function that used matrix, which we might not ever), but did in production, as it does eager loading of all code and tries to require `matrix` on boot, causing heroku deploy to fail.