Remove Rights and Repro func from *database*.
This needs to be deployed *after* #1722 has been deployed, in a two-step deploy, to avoid errors from heroku preboot meaning an old version of te code is still running after we've deleted db columns/tables it relies on.

So we'll leave this in "draft" PR until after we know #1722 has been deployed and taken effect, then merge and deploy again.