Refactor RightsTerms => RightsTerm
We have Rights Term information in a YAML file. Refactor the kind of odd functional API we had before to a more standard ruby model API, where there is a RightsTerm instance that has methods that give you things like .label etc.

There are still some class methods for convenience.

This will make it easier to design a custom rights page with where #1634 is evolving too.