Opt-in to attr_json feature for jsonb_contains query
attr_json gem provides this as an opt-in feature. Not sure why I left it out of kithe and out of here until now! It's super useful, makes it way easier to do queries on our json_attributes stuff without having to work it all out, including in situations with arrays.

https://github.com/jrochkind/attr_json#querying

For instance, ref #1751, wouldn't it be easier to do `Work.jsonb_contains("additional_credit.name": "Julie Huggony")`? Now we can.