make rights statement one big link
So icons is clickable. While remaining accessible -- it's harder to make links accessible if we have two links in a row pointing to the same thing one including an image etc (we don't really want a screen-reader to read two identical links in a row). Best to wrap it, I think.