rake task to fix hls_playlist_file_data double escaping
This is idempotent -- it won't harm the data if it's already been un-escaped. it just leaves it alone.

Also the OLD version of the attribute CAN read this fixed data, so you can fix the data first before new version of attribute is deployed. (The new version of the attribute can't read old data though)