upgrade Rails to 6.1.6.1
WITHOUT, for now, updating Blacklight. The slightly older version of Blacklight we are using requires some configuration to Rails, included here, to keep working with Rails 6.1.6.1. I don't believe the configuration is a security problem.

If we could update to latest Blacklight we would not need the custom configuration. And it woudl be a good idea anyway. But that requires updating to latest blacklight_range_limit too, which is currently causing a probelm in our app we need to figure out how to get around. In the meantime, let's update to the Rails with security patch. See #1781