DRY up check for can_see_unpublished_records?
Without changing semantics at all, we're just putting the same logic in a single method. Turns out you can put custom methods on your access_granted policy, so that's a convenient place to keep everything together -- but we also add an application helper for convenience, and actually delegate to that from the view components now. (At first i was _just_ going to make the application helper, but realized keeping access policy related logic in teh access policy class was possible, and nice)

Why do we want to DRY in one method? Cause this is kind of a weird call, hard to get right -- write it once right. And because we might want to change how this check for permission works, now one place to change it if we do.