restore robots.txt routing
It broke when we upgraded to Rails 7. Also added a test that was failing before I fixed, to hopefully avoid this again.

We are being kind of clever trying to route to robots.txt without an actual controller action method. Something about `text` vs `txt` broke it in Rails 7; possibly a bug in Rails 7 but I'm not going to spend the time to diagnose and report at present.

I started out by adding an explicit controller method, but getting _that_ to work forced me to see and deal with text vs txt issue, at which point I realized I didn't need the controller method if I switched the file suffix from `txt.erb` to `text.erb`. Don't really know why, but with test, good enough for now.