Fix review requested filter
Ref #2019

We were looking filtering for "not nil" but in fact our code left json `true` or `false`. We can use `jsonb_contains` to do the casting for us, we actually want to filter for records where value is `true` only, not `not nil`.

Still no tests, sorry, I'm being lazy. Tested manually.