Fix regexp, escaping hyphen
Prviously resulted in warning from ruby:

> /app/components/oral_history/transcript_component.rb:94: warning: character class has '-' without escape: /\A[[:space:]]*([A-Z-.\' ]+:) (.*)\Z/

This regex was copied from ohms viewer php, I think it is meant to mean A-Z (range), plus some punctuation including literal hyphen `-`, and this new escaped version is correct. Thanks ruby. Wonder what it was doing before and how it worked in php.