avoid trip to DB just to make sure home page work exists
We were previously doing a query to DB on every single home page view just to verify that the work 'w6634363x' really exists. What did we do if it DIDN'T? Just made a link that did nothing, instead of leading to a 404.  This isn't worth a trip to the DB on every home page view. If somehow that work gets deleted, the link will just be broken if you click on it, big deal, shouldn't happen, no big deal if it does, not worth a DB query on every single home page view.

Encountered while reviewing #2067